+++ flow-server/src/main/java/com/vaadin/ui/AngularTemplate.java
@@ -32,6 +32,7 @@ import com.vaadin.flow.nodefeature.TemplateMap;
 import com.vaadin.flow.router.HasChildView;
 import com.vaadin.flow.router.RouterConfiguration;
 import com.vaadin.flow.router.View;
+import com.vaadin.ui.event.AttachEvent;
 import com.vaadin.ui.polymertemplate.AbstractTemplate;
 import com.vaadin.flow.template.angular.RelativeFileResolver;
 import com.vaadin.flow.template.angular.TemplateNode;
@@ -42,7 +43,6 @@ import com.vaadin.flow.template.angular.model.TemplateModelProxyHandler;
 import com.vaadin.flow.template.angular.model.TemplateModelTypeParser;
 import com.vaadin.flow.template.angular.parser.TemplateParser;
 import com.vaadin.flow.template.angular.parser.TemplateResolver;
-import com.vaadin.ui.common.AttachEvent;
 import com.vaadin.util.ReflectTools;
 
 /**
