+++ flow-server/src/main/java/com/vaadin/ui/Component.java
@@ -24,17 +24,17 @@ import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.dom.ElementUtil;
 import com.vaadin.flow.dom.ShadowRoot;
 import com.vaadin.shared.Registration;
-import com.vaadin.ui.common.AttachEvent;
 import com.vaadin.ui.common.AttachNotifier;
-import com.vaadin.ui.common.DetachEvent;
 import com.vaadin.ui.common.DetachNotifier;
 import com.vaadin.ui.common.HasElement;
 import com.vaadin.ui.common.PropertyDescriptor;
 import com.vaadin.ui.common.PropertyDescriptors;
+import com.vaadin.ui.event.AttachEvent;
 import com.vaadin.ui.event.ComponentEvent;
 import com.vaadin.ui.event.ComponentEventBus;
 import com.vaadin.ui.event.ComponentEventListener;
 import com.vaadin.ui.event.ComponentEventNotifier;
+import com.vaadin.ui.event.DetachEvent;
 import com.vaadin.util.AnnotationReader;
 
 /**
