+++ flow-server/src/main/java/com/vaadin/ui/common/DetachNotifier.java
@@ -18,6 +18,7 @@ package com.vaadin.ui.common;
 import com.vaadin.ui.event.ComponentEventListener;
 import com.vaadin.shared.Registration;
 import com.vaadin.ui.event.ComponentEventNotifier;
+import com.vaadin.ui.event.DetachEvent;
 
 /**
  * Mixin interface for components that support adding detach listeners.
