+++ flow-tests/test-root-context/src/main/java/com/vaadin/flow/uitest/ui/DependencyView.java
@@ -27,9 +27,9 @@ import com.vaadin.ui.Component;
 import com.vaadin.ui.Tag;
 import com.vaadin.ui.Text;
 import com.vaadin.ui.UI;
-import com.vaadin.ui.common.AttachEvent;
 import com.vaadin.ui.common.HasText;
 import com.vaadin.ui.common.HtmlImport;
+import com.vaadin.ui.event.AttachEvent;
 import com.vaadin.ui.html.Div;
 import com.vaadin.ui.html.Hr;
 import com.vaadin.ui.html.NativeButton;
