+++ flow-tests/test-root-context/src/main/java/com/vaadin/flow/uitest/ui/InfoView.java
@@ -26,9 +26,9 @@ import com.vaadin.flow.router.View;
 import com.vaadin.function.DeploymentConfiguration;
 import com.vaadin.server.VaadinSession;
 import com.vaadin.server.WebBrowser;
-import com.vaadin.ui.common.AttachEvent;
 import com.vaadin.ui.Html;
 import com.vaadin.ui.UI;
+import com.vaadin.ui.event.AttachEvent;
 
 public class InfoView extends Div implements View {
