+++ flow-tests/test-root-context/src/main/java/com/vaadin/flow/uitest/ui/frontend/FrontendProtocolView.java
@@ -16,10 +16,10 @@
 package com.vaadin.flow.uitest.ui.frontend;
 
 import com.vaadin.flow.router.View;
-import com.vaadin.ui.common.AttachEvent;
 import com.vaadin.ui.Component;
 import com.vaadin.ui.Tag;
 import com.vaadin.ui.common.HasComponents;
+import com.vaadin.ui.event.AttachEvent;
 
 /**
  * Wrapper view for the {@link FrontendProtocolTemplate} component. This class
