+++ flow-tests/test-root-context/src/main/java/com/vaadin/flow/uitest/ui/template/PolymerModelPropertiesView.java
@@ -17,8 +17,8 @@ package com.vaadin.flow.uitest.ui.template;
 
 import com.vaadin.flow.router.View;
 import com.vaadin.ui.Tag;
-import com.vaadin.ui.common.AttachEvent;
 import com.vaadin.ui.common.HtmlImport;
+import com.vaadin.ui.event.AttachEvent;
 import com.vaadin.ui.event.ComponentEvent;
 import com.vaadin.ui.event.DomEvent;
 import com.vaadin.ui.html.Div;
